+++ CHANGELOG.md
@@ -2,3 +2,6 @@ Changelog is kept at [GitHub](https://github.com/Dronehub/minijson/releases),
 here's only the changelog for the version in development
 
 # v2.4
+
+* added argument default
+* fixing issue with serializing classes that subclass dict, list and tuple